Move account pages into pages directory
